Fix bug related to DO socket closure
Previously, I was aborting all pending fetch requests when a socket connection
with the partition DO was closed. This was a problem because multiple socket
connections can be open at the same time, and different pending fetch requests
can belong to different sockets. To fix this, when a socket connection is closed
only pending fetch requests belonging to that socket should be aborted.